+++ controllers/disruption_controller.go
@@ -29,12 +29,14 @@ import (
 	"math/rand"
 	"os"
 	"reflect"
+	"strings"
 	"time"
 
 	"github.com/cenkalti/backoff"
 	"go.uber.org/zap"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -46,6 +48,7 @@ import (
 
 	chaosapi "github.com/DataDog/chaos-controller/api"
 	chaosv1beta1 "github.com/DataDog/chaos-controller/api/v1beta1"
+	"github.com/DataDog/chaos-controller/env"
 	"github.com/DataDog/chaos-controller/metrics"
 	chaostypes "github.com/DataDog/chaos-controller/types"
 )
@@ -67,8 +70,8 @@ type DisruptionReconciler struct {
 	Recorder        record.EventRecorder
 	MetricsSink     metrics.Sink
 	DeleteOnly      bool
-	PodTemplateSpec corev1.Pod
 	TargetSelector  TargetSelector
+  InjectorAnnotations map[string]string
 	log             *zap.SugaredLogger
 }
 
@@ -279,7 +282,9 @@ func (r *DisruptionReconciler) startInjection(instance *chaosv1beta1.Disruption)
 	r.log.Infow("starting targets injection", "targets", instance.Status.Targets)
 
 	for _, target := range instance.Status.Targets {
-		var targetNodeName, containerID string
+		var targetNodeName string
+
+		var containerIDs []string
 
 		chaosPods := []*corev1.Pod{}
 
@@ -294,8 +299,8 @@ func (r *DisruptionReconciler) startInjection(instance *chaosv1beta1.Disruption)
 
 			targetNodeName = pod.Spec.NodeName
 
-			// get ID of targeted container or the first container
-			containerID, err = getContainerID(&pod, instance.Spec.Container)
+			// get IDs of targeted containers or all containers
+			containerIDs, err = getContainerIDs(&pod, instance.Spec.Containers)
 			if err != nil {
 				return fmt.Errorf("error getting target pod container ID: %w", err)
 			}
@@ -304,11 +309,7 @@ func (r *DisruptionReconciler) startInjection(instance *chaosv1beta1.Disruption)
 		}
 
 		// generate injection pods specs
-		if err := r.generateChaosPods(instance, &chaosPods, target, targetNodeName, containerID); err != nil {
-			r.log.Errorw("error generating injection chaos pod for target, skipping it", "error", err, "target", target)
-
-			continue
-		}
+		r.generateChaosPods(instance, &chaosPods, target, targetNodeName, containerIDs)
 
 		if len(chaosPods) == 0 {
 			r.Recorder.Event(instance, "Warning", "Empty Disruption", fmt.Sprintf("No disruption recognized for \"%s\" therefore no disruption applied.", instance.Name))
@@ -427,7 +428,7 @@ func (r *DisruptionReconciler) cleanDisruption(instance *chaosv1beta1.Disruption
 		// ignore it if it is not ready anymore
 		err := r.TargetSelector.TargetIsHealthy(target, r.Client, instance)
 		if err != nil {
-			if errors.IsNotFound(err) || err.Error() == "Pod is not Running" || err.Error() == "Node is not Ready" {
+			if errors.IsNotFound(err) || strings.ToLower(err.Error()) == "pod is not running" || strings.ToLower(err.Error()) == "node is not ready" {
 				// if the target is not in a good shape, we still run the cleanup phase but we don't check for any issues happening during
 				// the cleanup to avoid blocking the disruption deletion for nothing
 				r.log.Infow("target is not likely to be cleaned (either it does not exist anymore or it is not ready), the injector will TRY to clean it but will not take care about any failures", "target", target)
@@ -437,6 +438,7 @@ func (r *DisruptionReconciler) cleanDisruption(instance *chaosv1beta1.Disruption
 				ignoreCleanupStatus = true
 			} else {
 				r.log.Error(err.Error())
+				cleaned = false
 
 				continue
 			}
@@ -637,51 +639,186 @@ func (r *DisruptionReconciler) getChaosPods(instance *chaosv1beta1.Disruption, l
 
 // generatePod generates a pod from a generic pod template in the same namespace
 // and on the same node as the given pod
-func (r *DisruptionReconciler) generatePod(instance *chaosv1beta1.Disruption, targetName string, targetNodeName string, args []string, kind chaostypes.DisruptionKind) (*corev1.Pod, error) {
-	pod := corev1.Pod{}
-
+func (r *DisruptionReconciler) generatePod(instance *chaosv1beta1.Disruption, targetName string, targetNodeName string, args []string, kind chaostypes.DisruptionKind) *corev1.Pod {
 	image, ok := os.LookupEnv("CHAOS_INJECTOR_IMAGE")
 	if !ok {
 		image = "chaos-injector"
 	}
 
-	// copy pod template
-	data, err := json.Marshal(r.PodTemplateSpec)
-	if err != nil {
-		return nil, fmt.Errorf("error marshaling chaos pod template: %w", err)
-	}
+	// volume host path type definitions
+	hostPathDirectory := corev1.HostPathDirectory
+	hostPathFile := corev1.HostPathFile
 
-	err = json.Unmarshal(data, &pod)
-	if err != nil {
-		return nil, fmt.Errorf("error unmarshaling chaos pod template: %w", err)
-	}
-
-	// customize pod template
-	if pod.ObjectMeta.Labels == nil {
-		pod.ObjectMeta.Labels = map[string]string{}
-	}
-
-	pod.ObjectMeta.GenerateName = fmt.Sprintf("chaos-%s-", instance.Name)
-	pod.ObjectMeta.Namespace = instance.Namespace
-	pod.ObjectMeta.Labels[chaostypes.TargetLabel] = targetName
-	pod.ObjectMeta.Labels[chaostypes.DisruptionKindLabel] = string(kind)
-	pod.ObjectMeta.Labels[chaostypes.DisruptionNameLabel] = instance.Name
-	pod.Spec.NodeName = targetNodeName
-	pod.Spec.Containers[0].Image = image
-	pod.Spec.Containers[0].Args = args
-	pod.Spec.Containers[0].Env = append(pod.Spec.Containers[0].Env, corev1.EnvVar{
-		Name: chaostypes.TargetPodHostIPEnv,
-		ValueFrom: &corev1.EnvVarSource{
-			FieldRef: &corev1.ObjectFieldSelector{
-				FieldPath: "status.hostIP",
+	// define injector pod
+	pod := corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			GenerateName: fmt.Sprintf("chaos-%s-", instance.Name), // generate the pod name automatically with a prefix
+			Namespace:    instance.Namespace,                      // use instance namespace
+			Annotations:  r.InjectorAnnotations,                   // add extra annotations passed to the controller
+			Labels: map[string]string{
+				chaostypes.TargetLabel:         targetName,    // target name label
+				chaostypes.DisruptionKindLabel: string(kind),  // disruption kind label
+				chaostypes.DisruptionNameLabel: instance.Name, // disruption name label
 			},
 		},
-	})
+		Spec: corev1.PodSpec{
+			HostPID:       true,                      // enable host pid
+			RestartPolicy: corev1.RestartPolicyNever, // do not restart the pod on fail or completion
+			NodeName:      targetNodeName,            // specify node name to schedule the pod
+			Containers: []corev1.Container{
+				{
+					Name:            "injector",              // container name
+					Image:           image,                   // container image gathered from the environment variable
+					ImagePullPolicy: corev1.PullIfNotPresent, // pull the image only when it is not present
+					Args:            args,                    // pass disruption arguments
+					SecurityContext: &corev1.SecurityContext{
+						Privileged: func() *bool { b := true; return &b }(), // enable privileged mode
+					},
+					ReadinessProbe: &corev1.Probe{ // define readiness probe (file created by the injector when the injection is successful)
+						PeriodSeconds:    1,
+						FailureThreshold: 5,
+						Handler: corev1.Handler{
+							Exec: &corev1.ExecAction{
+								Command: []string{"cat", "/tmp/readiness_probe"},
+							},
+						},
+					},
+					Resources: corev1.ResourceRequirements{ // set resources requests and limits to zero
+						Limits: corev1.ResourceList{
+							corev1.ResourceCPU:    *resource.NewQuantity(0, resource.DecimalSI),
+							corev1.ResourceMemory: *resource.NewQuantity(0, resource.DecimalSI),
+						},
+						Requests: corev1.ResourceList{
+							corev1.ResourceCPU:    *resource.NewQuantity(0, resource.DecimalSI),
+							corev1.ResourceMemory: *resource.NewQuantity(0, resource.DecimalSI),
+						},
+					},
+					Env: []corev1.EnvVar{ // define environment variables
+						{
+							Name: env.InjectorTargetPodHostIP,
+							ValueFrom: &corev1.EnvVarSource{
+								FieldRef: &corev1.ObjectFieldSelector{
+									FieldPath: "status.hostIP",
+								},
+							},
+						},
+						{
+							Name: env.InjectorChaosPodIP,
+							ValueFrom: &corev1.EnvVarSource{
+								FieldRef: &corev1.ObjectFieldSelector{
+									FieldPath: "status.podIP",
+								},
+							},
+						},
+						{
+							Name:  env.InjectorMountHost,
+							Value: "/mnt/host/",
+						},
+						{
+							Name:  env.InjectorMountProc,
+							Value: "/mnt/host/proc/",
+						},
+						{
+							Name:  env.InjectorMountSysrq,
+							Value: "/mnt/sysrq",
+						},
+						{
+							Name:  env.InjectorMountSysrqTrigger,
+							Value: "/mnt/sysrq-trigger",
+						},
+						{
+							Name:  env.InjectorMountCgroup,
+							Value: "/mnt/cgroup/",
+						},
+					},
+					VolumeMounts: []corev1.VolumeMount{ // define volume mounts required for disruptions to work
+						{
+							Name:      "run",
+							MountPath: "/run",
+						},
+						{
+							Name:      "sysrq",
+							MountPath: "/mnt/sysrq",
+						},
+						{
+							Name:      "sysrq-trigger",
+							MountPath: "/mnt/sysrq-trigger",
+						},
+						{
+							Name:      "cgroup",
+							MountPath: "/mnt/cgroup",
+						},
+						{
+							Name:      "host",
+							MountPath: "/mnt/host",
+							ReadOnly:  true,
+						},
+					},
+				},
+			},
+			Volumes: []corev1.Volume{ // declare volumes required for disruptions to work
+				{
+					Name: "run",
+					VolumeSource: corev1.VolumeSource{
+						HostPath: &corev1.HostPathVolumeSource{
+							Path: "/run",
+							Type: &hostPathDirectory,
+						},
+					},
+				},
+				{
+					Name: "proc",
+					VolumeSource: corev1.VolumeSource{
+						HostPath: &corev1.HostPathVolumeSource{
+							Path: "/proc",
+							Type: &hostPathDirectory,
+						},
+					},
+				},
+				{
+					Name: "sysrq",
+					VolumeSource: corev1.VolumeSource{
+						HostPath: &corev1.HostPathVolumeSource{
+							Path: "/proc/sys/kernel/sysrq",
+							Type: &hostPathFile,
+						},
+					},
+				},
+				{
+					Name: "sysrq-trigger",
+					VolumeSource: corev1.VolumeSource{
+						HostPath: &corev1.HostPathVolumeSource{
+							Path: "/proc/sysrq-trigger",
+							Type: &hostPathFile,
+						},
+					},
+				},
+				{
+					Name: "cgroup",
+					VolumeSource: corev1.VolumeSource{
+						HostPath: &corev1.HostPathVolumeSource{
+							Path: "/sys/fs/cgroup",
+							Type: &hostPathDirectory,
+						},
+					},
+				},
+				{
+					Name: "host",
+					VolumeSource: corev1.VolumeSource{
+						HostPath: &corev1.HostPathVolumeSource{
+							Path: "/",
+							Type: &hostPathDirectory,
+						},
+					},
+				},
+			},
+		},
+	}
 
 	// add finalizer to the pod so it is not deleted before we can control its exit status
 	controllerutil.AddFinalizer(&pod, chaosPodFinalizer)
 
-	return &pod, nil
+	return &pod
 }
 
 // handleMetricSinkError logs the given metric sink error if it is not nil
@@ -701,6 +838,8 @@ func (r *DisruptionReconciler) validateDisruptionSpec(instance *chaosv1beta1.Dis
 			validator = instance.Spec.NodeFailure
 		case chaostypes.DisruptionKindNetworkDisruption:
 			validator = instance.Spec.Network
+		case chaostypes.DisruptionKindDNSDisruption:
+			validator = instance.Spec.DNS
 		case chaostypes.DisruptionKindCPUPressure:
 			validator = instance.Spec.CPUPressure
 		case chaostypes.DisruptionKindDiskPressure:
@@ -723,7 +862,7 @@ func (r *DisruptionReconciler) validateDisruptionSpec(instance *chaosv1beta1.Dis
 }
 
 // generateChaosPods generates a chaos pod for the given instance and disruption kind if set
-func (r *DisruptionReconciler) generateChaosPods(instance *chaosv1beta1.Disruption, pods *[]*corev1.Pod, targetName string, targetNodeName string, containerID string) error {
+func (r *DisruptionReconciler) generateChaosPods(instance *chaosv1beta1.Disruption, pods *[]*corev1.Pod, targetName string, targetNodeName string, containerIDs []string) {
 	// generate chaos pods for each possible disruptions
 	for _, kind := range chaostypes.DisruptionKinds {
 		var generator chaosapi.DisruptionArgsGenerator
@@ -734,6 +873,8 @@ func (r *DisruptionReconciler) generateChaosPods(instance *chaosv1beta1.Disrupti
 			generator = instance.Spec.NodeFailure
 		case chaostypes.DisruptionKindNetworkDisruption:
 			generator = instance.Spec.Network
+		case chaostypes.DisruptionKindDNSDisruption:
+			generator = instance.Spec.DNS
 		case chaostypes.DisruptionKindCPUPressure:
 			generator = instance.Spec.CPUPressure
 		case chaostypes.DisruptionKindDiskPressure:
@@ -752,19 +893,11 @@ func (r *DisruptionReconciler) generateChaosPods(instance *chaosv1beta1.Disrupti
 		}
 
 		// generate args for pod
-		args := generator.GenerateArgs(level, containerID, r.MetricsSink.GetSinkName(), instance.Spec.DryRun)
-
-		// generate pod
-		pod, err := r.generatePod(instance, targetName, targetNodeName, args, kind)
-		if err != nil {
-			return err
-		}
+		args := generator.GenerateArgs(level, containerIDs, r.MetricsSink.GetSinkName(), instance.Spec.DryRun)
 
 		// append pod to chaos pods
-		*pods = append(*pods, pod)
+		*pods = append(*pods, r.generatePod(instance, targetName, targetNodeName, args, kind))
 	}
-
-	return nil
 }
 
 // recordEventOnTarget records an event on the given target which can be either a pod or a node depending on the given disruption level
@@ -802,15 +935,19 @@ func (r *DisruptionReconciler) SetupWithManager(mgr ctrl.Manager) error {
 		Complete(r)
 }
 
-// WatchStuckOnRemoval lists disruptions every minutes and increment the "stuck on removal" metric
-// for every disruptions stuck on removal
-func (r *DisruptionReconciler) WatchStuckOnRemoval() {
+// ReportMetrics reports some controller metrics every minute:
+// - stuck on removal disruptions count
+// - ongoing disruptions count
+func (r *DisruptionReconciler) ReportMetrics() {
 	for {
 		// wait for a minute
 		<-time.After(time.Minute)
 
+		// declare counters
+		stuckOnRemoval := 0
+		chaosPodsCount := 0
+
 		l := chaosv1beta1.DisruptionList{}
-		count := 0
 
 		// list disruptions
 		if err := r.Client.List(context.Background(), &l); err != nil {
@@ -821,17 +958,32 @@ func (r *DisruptionReconciler) WatchStuckOnRemoval() {
 		// check for stuck ones
 		for _, d := range l.Items {
 			if d.Status.IsStuckOnRemoval {
-				count++
+				stuckOnRemoval++
 
 				if err := r.MetricsSink.MetricStuckOnRemoval([]string{"name:" + d.Name, "namespace:" + d.Namespace}); err != nil {
 					r.log.Errorw("error sending stuck_on_removal metric", "error", err)
 				}
 			}
+
+			chaosPods, err := r.getChaosPods(&d, nil)
+			if err != nil {
+				r.log.Errorw("error sending pods.count metric", "error", err)
+			}
+
+			chaosPodsCount += len(chaosPods)
 		}
 
-		// send count metric
-		if err := r.MetricsSink.MetricStuckOnRemovalCount(float64(count)); err != nil {
+		// send metrics
+		if err := r.MetricsSink.MetricStuckOnRemovalCount(float64(stuckOnRemoval)); err != nil {
 			r.log.Errorw("error sending stuck_on_removal_count metric", "error", err)
+		}
+
+		if err := r.MetricsSink.MetricDisruptionsCount(float64(len(l.Items))); err != nil {
+			r.log.Errorw("error sending disruptions.count metric", "error", err)
+		}
+
+		if err := r.MetricsSink.MetricPodsCount(float64(chaosPodsCount)); err != nil {
+			r.log.Errorw("error sending pods.count metric", "error", err)
 		}
 	}
 }