+++ controllers/disruption_controller_test.go
@@ -22,6 +22,7 @@ package controllers
 
 import (
 	"fmt"
+	"strings"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -103,6 +104,28 @@ func expectChaosPod(instance *chaosv1beta1.Disruption, count int) error {
 	return nil
 }
 
+// expectChaosInjectors retrieves the list of created chaos pods and confirms
+// that the targeted containers are present
+func expectChaosInjectors(instance *chaosv1beta1.Disruption, count int) error {
+	l, err := listChaosPods(instance)
+	if err != nil {
+		return err
+	}
+	for _, p := range l.Items {
+		args := p.Spec.Containers[0].Args
+		for i, arg := range args {
+			if arg == "--containers-id" {
+				containers := strings.Split(args[i+1], ",")
+				if len(containers) != count {
+					return fmt.Errorf("incorrect number of targeted containers in spec")
+				}
+			}
+		}
+	}
+
+	return nil
+}
+
 var _ = Describe("Disruption Controller", func() {
 	var disruption *chaosv1beta1.Disruption
 
@@ -113,7 +136,8 @@ var _ = Describe("Disruption Controller", func() {
 				Namespace: "default",
 			},
 			Spec: chaosv1beta1.DisruptionSpec{
-				Selector: map[string]string{"foo": "bar"},
+				Selector:   map[string]string{"foo": "bar"},
+				Containers: []string{"ctn1", "ctn2"},
 				NodeFailure: &chaosv1beta1.NodeFailureSpec{
 					Shutdown: false,
 				},
@@ -133,6 +157,15 @@ var _ = Describe("Disruption Controller", func() {
 						ReadBytesPerSec: func() *int { i := int(1); return &i }(),
 					},
 				},
+				DNS: []chaosv1beta1.HostRecordPair{
+					{
+						Hostname: "ctn",
+						Record: chaosv1beta1.DNSRecord{
+							Type:  "A",
+							Value: "10.0.0.1, 10.0.0.2 , 10.0.0.3",
+						},
+					},
+				},
 			},
 		}
 	})
@@ -140,11 +173,23 @@ var _ = Describe("Disruption Controller", func() {
 	AfterEach(func() {
 		// delete disruption resource
 		_ = k8sClient.Delete(context.Background(), disruption)
+		Eventually(func() error { return expectChaosPod(disruption, 0) }, timeout).Should(Succeed())
+		Eventually(func() error { return k8sClient.Get(context.Background(), instanceKey, disruption) }, timeout).Should(MatchError("Disruption.chaos.datadoghq.com \"foo\" not found"))
 	})
 
 	JustBeforeEach(func() {
 		By("Creating disruption resource")
 		Expect(k8sClient.Create(context.Background(), disruption)).To(BeNil())
+	})
+
+	Context("afterEach should clean an undeleted disruption", func() {
+		BeforeEach(func() {
+			disruption.Spec.Count = &intstr.IntOrString{Type: intstr.String, StrVal: "100%"}
+		})
+
+		It("should leave a created disruption for afterEach to clean", func() {
+			Eventually(func() error { return k8sClient.Get(context.Background(), instanceKey, disruption) }, timeout).Should(Succeed())
+		})
 	})
 
 	Context("target all pods", func() {
@@ -167,7 +212,10 @@ var _ = Describe("Disruption Controller", func() {
 			}, timeout).Should(Succeed())
 
 			By("Ensuring that the chaos pods have been created")
-			Eventually(func() error { return expectChaosPod(disruption, 16) }, timeout).Should(Succeed())
+			Eventually(func() error { return expectChaosPod(disruption, 20) }, timeout).Should(Succeed())
+
+			By("Ensuring that the chaos pods have correct number of targeted containers")
+			Expect(expectChaosInjectors(disruption, 2)).To(BeNil())
 
 			By("Deleting the disruption resource")
 			Expect(k8sClient.Delete(context.Background(), disruption)).To(BeNil())
@@ -188,7 +236,10 @@ var _ = Describe("Disruption Controller", func() {
 
 		It("should target all the selected pods", func() {
 			By("Ensuring that the inject pod has been created")
-			Eventually(func() error { return expectChaosPod(disruption, 4) }, timeout).Should(Succeed())
+			Eventually(func() error { return expectChaosPod(disruption, 5) }, timeout).Should(Succeed())
+
+			By("Ensuring that the chaos pods have correct number of targeted containers")
+			Expect(expectChaosInjectors(disruption, 2)).To(BeNil())
 
 			By("Deleting the disruption resource")
 			Expect(k8sClient.Delete(context.Background(), disruption)).To(BeNil())
@@ -209,7 +260,10 @@ var _ = Describe("Disruption Controller", func() {
 
 		It("should target all the selected pods", func() {
 			By("Ensuring that the inject pod has been created")
-			Eventually(func() error { return expectChaosPod(disruption, 12) }, timeout).Should(Succeed())
+			Eventually(func() error { return expectChaosPod(disruption, 15) }, timeout).Should(Succeed())
+
+			By("Ensuring that the chaos pods have correct number of targeted containers")
+			Expect(expectChaosInjectors(disruption, 2)).To(BeNil())
 
 			By("Deleting the disruption resource")
 			Expect(k8sClient.Delete(context.Background(), disruption)).To(BeNil())
@@ -219,6 +273,56 @@ var _ = Describe("Disruption Controller", func() {
 			Eventually(func() error { return expectChaosPod(disruption, 0) }, timeout).Should(Succeed())
 
 			By("Waiting for disruption resource to be deleted")
+			Eventually(func() error { return k8sClient.Get(context.Background(), instanceKey, disruption) }, timeout).Should(MatchError("Disruption.chaos.datadoghq.com \"foo\" not found"))
+		})
+	})
+
+	Context("target all pods and all containers by default", func() {
+		BeforeEach(func() {
+			disruption.Spec.Count = &intstr.IntOrString{Type: intstr.String, StrVal: "100%"}
+			disruption.Spec.Containers = []string{}
+		})
+
+		It("should target all the selected pods", func() {
+			By("Ensuring that the chaos pods have been created")
+			Eventually(func() error { return expectChaosPod(disruption, 16) }, timeout).Should(Succeed())
+
+			By("Ensuring that the chaos pods have correct number of targeted containers")
+			Expect(expectChaosInjectors(disruption, 3)).To(BeNil())
+
+			By("Deleting the disruption resource")
+			Expect(k8sClient.Delete(context.Background(), disruption)).To(BeNil())
+			Eventually(func() error { return k8sClient.Get(context.Background(), instanceKey, disruption) }, timeout).Should(Succeed())
+
+			By("Ensuring that the chaos pods have been deleted")
+			Eventually(func() error { return expectChaosPod(disruption, 0) }, timeout).Should(Succeed())
+
+			By("Waiting for disruption to be removed")
+			Eventually(func() error { return k8sClient.Get(context.Background(), instanceKey, disruption) }, timeout).Should(MatchError("Disruption.chaos.datadoghq.com \"foo\" not found"))
+		})
+	})
+
+	Context("target all pods and only one container is selected", func() {
+		BeforeEach(func() {
+			disruption.Spec.Count = &intstr.IntOrString{Type: intstr.String, StrVal: "100%"}
+			disruption.Spec.Containers = []string{"ctn1"}
+		})
+
+		It("should target all the selected pods", func() {
+			By("Ensuring that the chaos pods have been created")
+			Eventually(func() error { return expectChaosPod(disruption, 16) }, timeout).Should(Succeed())
+
+			By("Ensuring that the chaos pods have correct number of targeted containers")
+			Expect(expectChaosInjectors(disruption, 1)).To(BeNil())
+
+			By("Deleting the disruption resource")
+			Expect(k8sClient.Delete(context.Background(), disruption)).To(BeNil())
+			Eventually(func() error { return k8sClient.Get(context.Background(), instanceKey, disruption) }, timeout).Should(Succeed())
+
+			By("Ensuring that the chaos pods have been deleted")
+			Eventually(func() error { return expectChaosPod(disruption, 0) }, timeout).Should(Succeed())
+
+			By("Waiting for disruption to be removed")
 			Eventually(func() error { return k8sClient.Get(context.Background(), instanceKey, disruption) }, timeout).Should(MatchError("Disruption.chaos.datadoghq.com \"foo\" not found"))
 		})
 	})