+++ injector/network_disruption.go
@@ -12,6 +12,7 @@ import (
 	"time"
 
 	"github.com/DataDog/chaos-controller/api/v1beta1"
+	"github.com/DataDog/chaos-controller/env"
 	"github.com/DataDog/chaos-controller/network"
 	chaostypes "github.com/DataDog/chaos-controller/types"
 )
@@ -82,11 +83,15 @@ func (i networkDisruptionInjector) Inject() error {
 		i.addOutputLimitOperation(uint(i.spec.BandwidthLimit))
 	}
 
-	if err := i.applyOperations(); err != nil {
-		return fmt.Errorf("error applying tc operations: %w", err)
+	// apply operations if any
+	if len(i.operations) > 0 {
+		if err := i.applyOperations(); err != nil {
+			return fmt.Errorf("error applying tc operations: %w", err)
+		}
+
+		i.config.Log.Info("operations applied successfully")
 	}
 
-	i.config.Log.Info("operations applied successfully")
 	i.config.Log.Info("editing pod net_cls cgroup to apply a classid to target container packets")
 
 	// write classid to pod net_cls cgroup
@@ -229,9 +234,9 @@ func (i *networkDisruptionInjector) applyOperations() error {
 	i.config.Log.Infof("detected default gateway IP %s on interface %s", defaultRoute.Gateway().String(), defaultRoute.Link().Name())
 
 	// get the targeted pod node IP from the environment variable
-	hostIP, ok := os.LookupEnv(chaostypes.TargetPodHostIPEnv)
+	hostIP, ok := os.LookupEnv(env.InjectorTargetPodHostIP)
 	if !ok {
-		return fmt.Errorf("%s environment variable must be set with the target pod node IP", chaostypes.TargetPodHostIPEnv)
+		return fmt.Errorf("%s environment variable must be set with the target pod node IP", env.InjectorTargetPodHostIP)
 	}
 
 	i.config.Log.Infof("target pod node IP is %s", hostIP)