+++ types/types.go
@@ -17,8 +17,6 @@ type DisruptionInjectionStatus string
 const (
 	// TargetLabel is the label used to identify the pod targeted by a chaos pod
 	TargetLabel = "chaos.datadoghq.com/target"
-	// TargetPodHostIPEnv is the target pod host IP environment variable name
-	TargetPodHostIPEnv = "TARGET_POD_HOST_IP"
 
 	// DisruptionKindLabel is the label used to identify the disruption kind for a chaos pod
 	DisruptionKindLabel = "chaos.datadoghq.com/disruption-kind"
@@ -30,6 +28,8 @@ const (
 	DisruptionKindCPUPressure = "cpu-pressure"
 	// DisruptionKindDiskPressure is a disk pressure disruption
 	DisruptionKindDiskPressure = "disk-pressure"
+	// DisruptionKindDNSDisruption is a dns disruption
+	DisruptionKindDNSDisruption = "dns-disruption"
 
 	// DisruptionLevelUnspecified is the value used when the level of injection is not specified
 	DisruptionLevelUnspecified = ""
@@ -56,5 +56,6 @@ var (
 		DisruptionKindNodeFailure,
 		DisruptionKindCPUPressure,
 		DisruptionKindDiskPressure,
+		DisruptionKindDNSDisruption,
 	}
 )